Move logic of `HeaderProduct` to the new `ProductSwitch` component
