+++ src/components/ProductSwitch/index.ts
@@ -1,0 +1,1 @@
+export * from "./ProductSwitch";
