--- src/components/SwitchProductButton/index.ts
@@ -1,1 +1,0 @@
-export * from "./SwitchProductButton";
